Added logic for two-state mappings to be treated as normal buttons
This is primarily to get around issues related to using groups.
The inner item type is known as a group, which means it cannot
be treated as a button. In my case, this meant that the Hue lights
that were grouped together could not be used with the pebble app.